Fix: Accent colour reverts to default when leaving dash
